Intermediate: ergonomic view of the path (or generated path from folder)
matverse.vista allows a preliminary peak at what would be added to path
by matverse.enter

pathshow() prints the current MATLABPATH in a user-friendly way, namely
with splitted entries and reversed order, so that last added paths are
the first you see.

pathshow('custom_dir') generates a recursive path from 'custom_dir/',
and displays again splitted and reversed.

————————————————————————————————————————————————————————————————————————

We also add pathdef.m to the .gitignore. Such a file has the purpose of
letting the user save a path, whenever the global MATLAB folder is not
reachable (e.g. because is propriety of root and you are just a user...)
> We don't want to track those files, of course.