Intermediate: add startup.m to ignored git files
Apparently if you are not root, you cannot just savepath('pathdef.m'),
but you have also to load that file whenever you start a new session.

So we need to define a startup.m file in the userpath folder, containing

    addpath(pathdef);

Again we don't want to track these technical files.

TODO: write about it on the README.