Add preliminary support for GNU Octave
I have not really clear which features are missing in Octave, so this
would probably be a matter of intensive trial and error. For now I got
that Octave does not have access to private package subdirectories.

> solved partially with some if statements here and there (but without
  assuring the very same functionality, i.e. giving up xgenpath and co.)

> a better solution would be to define a +octaverse package, with the
  aim to maintain it seperately from +matverse (and so exclude all the
  other functions and packages that would never work with Octave, like
  e.g. cprintf...). This would require a lot of testing.

We provide also a simple function, within the scriptverse, to determine
if running under MATLAB or GNU Octave: isoctave() -> {true,false} :)